refactor(#1570): extract successor-offset helpers to partition_successor.rs (campsite, keep partition_lookup.rs <800)
